Entregables/JoseDavidBarritaLopez/AutomationFramework/Behave/sauce_labs/requirements.txt
allure-behave==2.8.6
allure-python-commons==2.8.6
attrs==19.2.0
behave==1.2.6
nose==1.3.7
Paver==1.3.4
selenium==3.141.0
zipp==0.6.0
typing-extensions==4.1.1
requests~=2.27.1
python-dotenv~=0.19.2
Appium-Python-Client==1.0.2
python-dateutil==2.8.2